make login async again
