+++ src/main/java/me/iblitzkriegi/vixio/effects/EffLogin.java
@@ -1,16 +1,11 @@
 package me.iblitzkriegi.vixio.effects;
 
-import ch.njol.skript.Skript;
-import ch.njol.skript.effects.Delay;
-import ch.njol.skript.lang.Effect;
 import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser;
-import ch.njol.skript.lang.TriggerItem;
 import ch.njol.util.Kleenean;
 import me.iblitzkriegi.vixio.Vixio;
 import me.iblitzkriegi.vixio.commands.CommandListener;
 import me.iblitzkriegi.vixio.events.base.EventListener;
-import me.iblitzkriegi.vixio.jda.JDAEventListener;
 import me.iblitzkriegi.vixio.util.AsyncEffect;
 import me.iblitzkriegi.vixio.util.wrapper.Bot;
 import net.dv8tion.jda.core.AccountType;
@@ -19,17 +14,10 @@ import net.dv8tion.jda.core.JDABuilder;
 import net.dv8tion.jda.core.exceptions.AccountTypeException;
 import org.bukkit.Bukkit;
 import org.bukkit.event.Event;
-import org.bukkit.scheduler.BukkitScheduler;
 
 import javax.security.auth.login.LoginException;
-import java.lang.reflect.Field;
-import java.util.Set;
 
-/**
- * This effect must be synchronous or else events may parse before it and their listeners
- * wont be properly registered to the new bot.
- */
-public class EffLogin extends Effect {
+public class EffLogin extends AsyncEffect {
 
     static {
         Vixio.getInstance().registerEffect(EffLogin.class, "(login|connect) to %string% (using|with) [the] name %string%")
@@ -48,12 +36,9 @@ public class EffLogin extends Effect {
         }
         if (Vixio.getInstance().botNameHashMap.get(name) != null) {
             // just to make the error show up outside of skript's reload errors
-            Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(Vixio.getInstance(), new Runnable() {
-                @Override
-                public void run() {
-                    Vixio.getErrorHandler().warn("Vixio attempted to login to a bot with the name " + name + " but a bot already exists with that name.");
-                }
-            }, 1);
+            Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(Vixio.getInstance(),
+                    () -> Vixio.getErrorHandler().warn("Vixio attempted to login to a bot with the name " + name + " but a bot already exists with that name."),
+                    1);
             return;
         }
         JDA api = null;
@@ -68,11 +53,14 @@ public class EffLogin extends Effect {
                 e1.printStackTrace();
             }
         }
+        // Make the new bot listen to active events and commands
+        for (EventListener<?> listener : EventListener.listeners) {
+            api.addEventListener(listener);
+        }
         api.addEventListener(new CommandListener());
         Bot bot = new Bot(name, api);
         Vixio.getInstance().botHashMap.put(api, bot);
         Vixio.getInstance().botNameHashMap.put(name, bot);
-
     }
 
     @Override
