--- src/main/java/me/iblitzkriegi/vixio/events/base/SimpleBukkitEvent.java
@@ -1,14 +1,0 @@
-package me.iblitzkriegi.vixio.events.base;
-
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
-
-public class SimpleBukkitEvent extends Event {
-
-    private HandlerList handlerList = new HandlerList();
-
-    public HandlerList getHandlers() {
-        return handlerList;
-    }
-
-}
